Move buildObjectMapper method to Json class
